+++ new-api/src/v2/modules/works/services/areaExpertise/createAreaExpertise.service.ts
@@ -1,6 +1,7 @@
 import { EntityMember } from '@modules/members/typeorm/entities/member.entity';
 import IRepositoryAreaExpertise from '@modules/works/repositories/IRepositoryAreaExpertise';
 import { ConflictException, ForbiddenException } from '@nestjs/common';
+import { Unauthorized } from 'v2/utils/Errors/Unauthorized';
 import ICreateAreaExpertiseDTO from '../../dtos/areaExpertise/ICreateAreaExpertise.dto';
 import { EntityAreaExpertise } from '../../typeorm/entities/areaExpertise.entity';
 
@@ -14,6 +15,9 @@ interface IRequest {
 const create = async (params: IRequest): Promise<EntityAreaExpertise> => {
   const { repository, data, member } = params;
   // Todo verificar a patente do membro
+  if (!member) {
+    throw new Unauthorized();
+  }
 
   const { patent } = member;
