+++ new-api/src/v2/modules/works/services/areaExpertise/index.ts
@@ -1,5 +1,4 @@
 import create from './createAreaExpertise.service';
 import findAll from './findAllAreaExpertise.service';
-import findByIdMember from './findMemberById.service';
 
-export { create, findAll, findByIdMember };
+export { create, findAll };
